Fix Bugs in Ubuntu 1304

+++ ZoomPipeline_FuncSvr/smartlink/st_clientnode.cpp
@@ -272,7 +272,7 @@ void st_clientNode::CheckHeartBeating()
     qint64 usc = this->m_last_Report.secsTo(dtm);
     if (usc >=60)
     {
-        emit evt_Message(tr("Client ") + QString("%1").arg((unsigned int)(this)) + tr(" is dead, kick out."));
+        emit evt_Message(tr("Client ") + QString("%1").arg((unsigned int)((quint64)this)) + tr(" is dead, kick out."));
         emit evt_close_client(this->sock());
     }
 }
